requirements.txt
albumentations
autogluon
av
ConfigSpace
dataclasses_json
datasets
decord
Flask
flask_cors
gpytorch
huggingface_hub
hydra-core
imagesize
iopath
kagglehub
matplotlib
mkdocs_gen_files
more_itertools
numpy
omegaconf
pandas
peft
Pillow
pycocotools
PyYAML
rarfile
regex
Requests
scikit_learn
scipy
segmentation_models_pytorch
setuptools
skimage
strawberry
submitit
tensordict
torch
torchmetrics
torchvision
tqdm
transformers
typing_extensions